add sweet alert

+++ src/components/team/members/Professional.jsx
@@ -2,6 +2,7 @@ import { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { useParams, useNavigate } from 'react-router-dom';
 import { GiPlagueDoctorProfile } from 'react-icons/gi';
+import Swal from 'sweetalert2';
 import { getProfessionalById, getAllProfessionals, deleteProfessional } from '../../../actions/teams/thunksteam';
 
 const Professional = () => {
@@ -14,14 +15,28 @@ const Professional = () => {
     navigate(`/edit_professional/${id}`);
   };
 
-  const handleDelete = async () => {
-    try {
-      await dispatch(deleteProfessional(id));
-      await dispatch(getAllProfessionals());
-      navigate('/team');
-    } catch (error) {
-      throw new Error(error);
-    }
+  const handleDelete = (id) => {
+    Swal.fire({
+      title: 'Delete Item',
+      text: 'Are you sure you want to delete this professional?',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#d33',
+      cancelButtonColor: '#3085d6',
+      confirmButtonText: 'Yes, delete it!',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        dispatch(deleteProfessional(id))
+          .then(() => {
+            dispatch(getAllProfessionals());
+            Swal.fire('Deleted!', 'professional has been deleted.', 'success');
+            navigate('/team');
+          })
+          .catch((error) => {
+            Swal.fire('Error', 'An error occurred while deleting the item.', error);
+          });
+      }
+    });
   };
 
   useEffect(() => {
